% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assess_impact_absolute_risk.R
\name{assess_impact_absolute_risk}
\alias{assess_impact_absolute_risk}
\title{input_withPaf of Health Impacts}
\usage{
assess_impact_absolute_risk(
  exp,
  pop_exp,
  erf_shape,
  erf_parameters,
  info_pollutant = NULL,
  info_outcome = NULL,
  info_exp = NULL,
  info_cf = NULL,
  info_erf = NULL
)
}
\arguments{
\item{exp}{\code{Vector} showing the exposure categories (average of the exposure ranges) in a exposure distribution refering only to exposed population. The length of exp and pop_exp should be the same.}

\item{pop_exp}{\code{Numeric value} or {vector} showing the population exposed for each of the exposure categories. The length of this input variable should be the same as "exp".}

\item{erf_shape}{\code{String} to choose among "linear", "loglinear" and "quadratic".}

\item{erf_parameters}{\code{Vector} of numeric values as in order of apereance in the exposure response function.}

\item{info_pollutant}{\code{String} showing additional information or id for the pollutant. Default value = NULL.}

\item{info_outcome}{\code{String} showing additional information or id for the health outcome. Default value = NULL.}

\item{info_exp}{\code{String} showing additional information or id for the exposure. This information will be added to all rows of the results. Default value = NULL.}

\item{info_cf}{\code{String} showing additional information or id for counter-factual scenario (cut-off). This information will be added to all rows of the results. Default value = NULL.}

\item{info_ar}{\code{String} showing additional information or id for the concentration-response function. This information will be added to all rows of the results. Default value = NULL.}
}
\value{
This function returns a \code{data.frame} with one row for each value of the
concentration-response function (i.e. mean, lower and upper bound confidence interval.
Moreover, the data frame include columns such as:
\itemize{
 \item Attributable fraction
 \item Health impact
 \item Outcome metric
 \item And many more.
}
}
\description{
Calculates the health impacts, mortality or morbidity, of an environmental stressor using a single value for baseline heath data, i.e. without life table. It provides as a result the mean as well as the lower and the higher bound of the impact based on the confidence interval of the concentration-response function.
}
\note{
Experimental function
}
\examples{
TBD
}
\author{
Alberto Castro
}
