% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assess_mortality_lifetable.R
\name{assess_mortality_lifetable}
\alias{assess_mortality_lifetable}
\title{Health impacts based on life tables}
\usage{
assess_mortality_lifetable(
  exp,
  prop_pop_exp = 1,
  cf,
  crf,
  crf_per,
  crf_rescale_method,
  first_age_pop,
  last_age_pop,
  interval_age_pop,
  prob_natural_death_male,
  prob_natural_death_female,
  prob_total_death_male,
  prob_total_death_female,
  population_male,
  population_female,
  year_of_analysis,
  corrected_discount_rate = 0,
  min_age = NULL,
  max_age = NULL,
  info_pollutant = NULL,
  info_outcome = NULL,
  info_exp = NULL,
  info_cf = NULL,
  info_crf = NULL,
  info_bhd = NULL
)
}
\arguments{
\item{exp}{\code{Numeric value} showing the population-weighted mean exposure in ug/m3 or {vector} showing the exposure category in a exposure distribution (this information is linked to the proportion of population exposed).}

\item{prop_pop_exp}{\code{Numeric value} or {vector} showing the proportion of population exposed (as fraction, i.e. values between 0 and 1) for a single exposure value or for multiple categories, i.e., a exposure distribution, respectively. If a exposure distribution is used, the dimension of this input variable should be the same as "exp". By default, 1 for single exposure value will be assigned to this input variable assuming a single exposure value, but users can change this value.}

\item{crf}{\code{Vector} of three numeric values referring to the mean as well as the lower bound and upper bound of the confidence interval.}

\item{crf_per}{\code{Numberic value} showing the increment of the concentration-response function in ug/m3 (usually 10 or 5).}

\item{crf_rescale_method}{\code{String} to choose among "linear" and "loglinear".}

\item{first_age_pop}{\code{Numberic value} of the first item of the age sequence from population and life table data}

\item{last_age_pop}{\code{Numberic value} of the last item of the age sequence from population and life table data}

\item{interval_age_pop}{\code{Numberic value} of the interval of the age sequence from population and life table data}

\item{prob_natural_death_male}{\code{Vector} containing the probability of dying because of natural cause (excluding non-natural deaths such as violence or accidents) by age or age group for male.}

\item{prob_natural_death_female}{\code{Vector} containing the probability of dying because of natural cause (excluding non-natural deaths such as violence or accidents) by age or age group for female.}

\item{prob_total_death_male}{\code{Vector} containing the probability of dying because of all causes (including non-natural deaths such as violence or accidents) by age or age group for male.}

\item{prob_total_death_female}{\code{Vector} containing the probability of dying because of all causes (including non-natural deaths such as violence or accidents) by age or age group for female.}

\item{population_male}{\code{Vector} containing the mid-year male population for the year of analysis.}

\item{population_female}{\code{Vector} containing the mid-year female population for the year of analysis.}

\item{year_of_analysis}{Numeric value of the year of analysis, which corresponds to the first year of the life table.}

\item{corrected_discount_rate}{Numeric value with the corrected discount rate as proportion (i.e. 0.1 instead of 10\%).}

\item{min_age}{\code{Numberic value} of the minimal age to be considered for adults (by default 30, i.e. 30+).}

\item{max_age}{\code{Numberic value} of the maximal age to be considered for infants/children (by default 0, i.e. below 1 years old).}

\item{info_pollutant}{\code{String} showing additional information or id for the pollutant. Default value = NULL.}

\item{info_outcome}{\code{String} showing additional information or id for the health outcome. Default value = NULL.}

\item{info_exp}{\code{String} showing additional information or id for the exposure. This information will be added to all rows of the results. Default value = NULL.}

\item{info_cf}{\code{String} showing additional information or id for counter-factual scenario (cut-off). This information will be added to all rows of the results. Default value = NULL.}

\item{info_crf}{\code{String} showing additional information or id for the concentration-response function. This information will be added to all rows of the results. Default value = NULL.}

\item{info_bhd}{\code{String} showing additional information or id for the baseline health data. This information will be added to all rows of the results. Default value = NULL.}
}
\value{
This function returns a \code{data.frame} with one row for each value of the
concentration-response function (i.e. mean, lower and upper bound confidence interval.
Moreover, the data frame include columns such as:
\itemize{
 \item Attributable fraction
 \item Health impact
 \item Outcome metric
 \item And many more.
}
}
\description{
Calculates the mortality, i.e. premature deaths or years of life lost, attributed to the exposure to an environmental stressor using a life table approach. It provides as a result the mean as well as the lower and the higher bound of the impact based on the confidence interval of the concentration-response function.
}
\note{
Experimental function
}
\examples{
TBD
}
\author{
Alberto Castro
}
