% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attribute_copy.R
\name{attribute_copy}
\alias{attribute_copy}
\title{Attributable health cases based on relative risk}
\usage{
attribute_copy(
  health_metric = "same_input_output",
  risk_method = "relative_risk",
  exp_central,
  exp_lower = NULL,
  exp_upper = NULL,
  prop_pop_exp = 1,
  pop_exp = NULL,
  cutoff,
  rr_central = NULL,
  rr_lower = NULL,
  rr_upper = NULL,
  erf_increment = NULL,
  erf_shape = NULL,
  erf_c_central = NULL,
  erf_c_lower = NULL,
  erf_c_upper = NULL,
  bhd_central = NULL,
  bhd_lower = NULL,
  bhd_upper = NULL,
  disability_weight = NULL,
  duration = NULL,
  first_age_pop = NULL,
  last_age_pop = NULL,
  prob_natural_death_male = NULL,
  prob_natural_death_female = NULL,
  prob_total_death_male,
  prob_total_death_female = NULL,
  population_midyear_male = NULL,
  population_midyear_female = NULL,
  year_of_analysis = NULL,
  min_age = NULL,
  max_age = NULL,
  corrected_discount_rate = NULL,
  geo_id_raw = NULL,
  geo_id_aggregated = NULL,
  info = NULL
)
}
\arguments{
\item{risk_method}{\code{String} showing the calculation method of the risk: "relative_risk" or "absolute_risk".}

\item{exp_central, exp_lower, exp_upper}{\code{Numeric values} of the exposure
to the environmental stressor referring to the central estimate and (optionally)
to lower and upper bound of the confidence interval. If only one value is provided,
it will be assumed that it refers to population-weighted mean exposure in ug/m3.
If a {vector} is provided, it will be assumed that it refers to the exposure categories
(average exposure in the category) in a exposure distribution
(this information is linked to the proportion of population exposed).}

\item{prop_pop_exp}{\code{Numeric value} or {Numeric vector} Fraction (values between 0 & 1) of the total population exposed to (one or more) exposure categories, i.e., a exposure distribution, respectively. If a exposure distribution is used, the dimension of this input variable should be the same as "exp". By default, 1 for single exposure value will be assigned to this input variable assuming a single exposure value, but users can change this value.}

\item{pop_exp}{\code{Numeric value} or {vector} showing the population exposed for each of the exposure categories. The length of this input variable must be the same as "exp".}

\item{cutoff}{\code{Numeric value} showing the cut-off exposure in ug/m3 (i.e. the exposure level below which no health effects occur).}

\item{rr_central, rr_lower, rr_upper}{\code Three {numeric values} referring to the central estimate of the relative risk and the corresponding lower and upper 95\% confidence interval bounds.}

\item{erf_increment}{\code{Numeric value} showing the increment of the exposure-response function in ug/m3 (usually 10 or 5).}

\item{erf_shape}{\code{String} showing the shape of the exposure-response function to be assumed using the relative risk from the literature as support point. Options: "linear", log_linear", "linear_log", "log_log".}

\item{bhd_central}{\code{Numeric value} showing the only or central estimate (if confidence interval) of the baseline health data (e.g. incidence of the health outcome in the population).}

\item{bhd_lower}{\code{Numeric value} showing the upper bound estimate (confidence interval) of the baseline health data.}

\item{bhd_upper}{\code{Numeric value} showing the lower bound estimate (confidence interval) of the baseline health data.}

\item{disability_weight}{\code{Numeric value} showing the disability weight associated with the morbidity health outcome}

\item{duration}{\code{Numeric value} showing the disease duration
estimate and (optionally) the lower bound and the upper bound of the confidence
interval of the baseline health data (e.g. incidence of the health outcome in the population).}

\item{first_age_pop}{\code{Numeric value} starting age of the youngest age group from population and life table data (age interval = 1 year)}

\item{last_age_pop}{\code{Numeric value} ending age of the oldest age group from population and life table data (age interval = 1 year)}

\item{prob_natural_death_male, prob_natural_death_female}{\code{Numeric vector} containing the probability of dying due to natural cause (excluding non-natural deaths due to violence or accidents) by age or age group for male and female respectively.}

\item{prob_total_death_male, prob_total_death_female}{\code{Numeric vector} containing the probability of dying due to all causes (including non-natural deaths due to violence or accidents) by age or age group for male and female respectively.}

\item{population_midyear_male, population_midyear_female}{\code{Numeric vector} containing the mid-year male population for the year of analysis for male and female respectively.}

\item{year_of_analysis}{\code{Numeric value} of the year of analysis, which corresponds to the first year of the life table.}

\item{min_age}{\code{Numberic value} of the minimal age to be considered for adults (by default 30, i.e. 30+).}

\item{max_age}{\code{Numberic value} of the maximal age to be considered for infants/children (by default 0, i.e. below 1 year old).}

\item{info}{\code{String} or {data frame} showing additional information or id. The suffix "info" will be added to the column name. Default value = NULL.}
}
\value{
TBD. E.g. This function returns a \code{data.frame} with one row for each value of the
concentration-response function (i.e. central, lower and upper bound confidence interval.
Moreover, the data frame includes columns such as:
\itemize{
 \item Attributable fraction
 \item Health impact
 \item Outcome metric
 \item And many more.
}
}
\description{
Calculates the health impacts, mortality or morbidity, of an environmental stressor using a single value for baseline heath data, i.e. without life table. It provides as a result the mean as well as the lower and the higher bound of the impact based on the confidence interval of the concentration-response function.
}
\examples{
TBD
}
\author{
Alberto Castro
}
