% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attribute_deaths_lifetable_rr.R
\name{attribute_deaths_lifetable_rr}
\alias{attribute_deaths_lifetable_rr}
\title{Attributable deaths based on life tables}
\usage{
attribute_deaths_lifetable_rr(
  exp,
  prop_pop_exp = 1,
  cutoff,
  rr,
  rr_increment,
  erf_shape,
  erf_c = NULL,
  first_age_pop,
  last_age_pop,
  interval_age_pop,
  prob_natural_death_male,
  prob_natural_death_female,
  prob_total_death_male,
  prob_total_death_female,
  population_male,
  population_female,
  year_of_analysis,
  min_age = NULL,
  max_age = NULL,
  info = NULL
)
}
\arguments{
\item{exp}{\code{Numeric values} Population-weighted mean exposure in ug/m3 or {vector} showing the exposure category in a exposure distribution (this information is linked to the proportion of population exposed).}

\item{prop_pop_exp}{\code{Numeric value} or {Numeric vector} Fraction (values between 0 & 1) of the total population exposed to (one or more) exposure categories, i.e., a exposure distribution, respectively. If a exposure distribution is used, the dimension of this input variable should be the same as "exp". By default, 1 for single exposure value will be assigned to this input variable assuming a single exposure value, but users can change this value.}

\item{cutoff}{\code{Numeric value} showing the cut-off exposure in ug/m3 (i.e. the exposure level below which no health effects occur).}

\item{rr}{\code{Numeric vector} of three numeric values referring to the central estimate of the exposure-response function and the corresponding lower and upper 95\% confidence interval bounds.}

\item{rr_increment}{\code{Numeric value} showing the increment of the exposure-response function in ug/m3 (usually 10 or 5).}

\item{erf_shape}{\code{String} showing the shape of the exposure-response function to be assumed using the relative risk from the literature as support point. Options: "linear", log_linear", "linear_log", "log_log".}

\item{erf_c}{\code{String} showing the user-defined function that puts the relative risk in relation with concentration. The function must have only one variable: c, which means concentration. E.g. "3+c+c^2". Default value = NULL.}

\item{first_age_pop}{\code{Numeric value} starting age of the youngest age group from population and life table data}

\item{last_age_pop}{\code{Numeric value} ending age of the oldest age group from population and life table data}

\item{interval_age_pop}{\code{Numeric value} of the interval (in years) of each age group from population and life table data}

\item{prob_natural_death_male}{\code{Numeric vector} containing the probability of dying due to natural cause (excluding non-natural deaths due to violence or accidents) by age or age group for males.}

\item{prob_natural_death_female}{\code{Numeric vector} containing the probability of dying due to natural cause (excluding non-natural deaths due to violence or accidents) by age or age group for females.}

\item{prob_total_death_male}{\code{Numeric vector} containing the probability of dying due to all causes (including non-natural deaths due to violence or accidents) by age or age group for males.}

\item{prob_total_death_female}{\code{Numeric vector} containing the probability of dying due to all causes (including non-natural deaths due to violence or accidents) by age or age group for females.}

\item{population_male}{\code{Numeric vector} containing the mid-year male population for the year of analysis.}

\item{population_female}{\code{Vector} containing the mid-year female population for the year of analysis.}

\item{year_of_analysis}{\code{Numeric value} of the year of analysis, which corresponds to the first year of the life table.}

\item{min_age}{\code{Numberic value} of the minimal age to be considered for adults (by default 30, i.e. 30+).}

\item{max_age}{\code{Numberic value} of the maximal age to be considered for infants/children (by default 0, i.e. below 1 year old).}

\item{info}{\code{String} showing additional information or id for the pollutant. The suffix "info" will be added to the column name. Default value = NULL.}
}
\value{
This function returns a \code{data.frame} with one row for each value of the
concentration-response function (i.e. mean, lower and upper bound confidence interval.
Moreover, the data frame include columns such as:
\itemize{
 \item Attributable fraction
 \item Health impact
 \item Outcome metric
 \item And many more.
}
}
\description{
Calculates the premature deaths attributable to the exposure to an environmental stressor using a life table approach. It provides the central estimate of the impact and the corresponding 95\% confidence intervals (based on the 95\% confidence interval exposure-response function).
}
\note{
Experimental function
}
\examples{
TBD
}
\author{
Alberto Castro
}
