% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attribute_health.R
\name{attribute_health}
\alias{attribute_health}
\title{Attribute health impact from the same health metric and using on relative risk}
\usage{
attribute_health(
  exp_central,
  exp_lower = NULL,
  exp_upper = NULL,
  prop_pop_exp = 1,
  cutoff,
  rr_central = NULL,
  rr_lower = NULL,
  rr_upper = NULL,
  erf_increment = NULL,
  erf_shape = NULL,
  erf_c_central = NULL,
  erf_c_lower = NULL,
  erf_c_upper = NULL,
  bhd_central,
  bhd_lower = NULL,
  bhd_upper = NULL,
  geo_id_raw = NULL,
  geo_id_aggregated = NULL,
  info = NULL
)
}
\arguments{
\item{prop_pop_exp}{\code{Numeric value} or {Numeric vector} Fraction (values between 0 & 1) of the total population exposed to (one or more) exposure categories, i.e., a exposure distribution, respectively. If a exposure distribution is used, the dimension of this input variable should be the same as "exp". By default, 1 for single exposure value will be assigned to this input variable assuming a single exposure value, but users can change this value.}

\item{cutoff}{\code{Numeric value} showing the cut-off exposure in ug/m3 (i.e. the exposure level below which no health effects occur).}

\item{erf_increment}{\code{Numeric value} showing the increment of the exposure-response function in ug/m3 (usually 10 or 5).}

\item{erf_shape}{\code{String} showing the shape of the exposure-response function to be assumed using the relative risk from the literature as support point. Options: "linear", log_linear", "linear_log", "log_log".}

\item{bhd_central, bhd_lower, bhd_upper}{\code{Numeric value} showing the central
estimate and (optionally) the lower bound and the upper bound of the confidence
interval of the baseline health data (e.g. incidence of the health outcome in the population).}

\item{info}{\code{String} or {data frame} showing additional information or id. The suffix "info" will be added to the column name. Default value = NULL.}
}
\value{
TBD. E.g. This function returns a \code{data.frame} with one row for each value of the
concentration-response function i.e. central, lower and upper bound confidence interval.
Moreover, the data frame includes columns such as:
\itemize{
 \item Attributable fraction
 \item Health impact
 \item Outcome metric
 \item And many more.
}
}
\description{
It calculates the health impacts, mortality or morbidity,
of exposure to an environmental stressor based on the same health metric
(in the baseline health data as in the result) and using relative risk.
}
\note{
Experimental function
}
\examples{
TBD
}
\author{
Alberto Castro
}
