% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attribute_health_ar.R
\name{attribute_health_ar}
\alias{attribute_health_ar}
\title{Attributable health cases based on absolute risk}
\usage{
attribute_health_ar(
  exp_central,
  exp_lower = NULL,
  exp_upper = NULL,
  pop_exp,
  erf_c_central,
  erf_c_lower = NULL,
  erf_c_upper = NULL,
  geo_id_raw = NULL,
  geo_id_aggregated = NULL,
  info = NULL
)
}
\arguments{
\item{exp_central, exp_lower, exp_upper}{\code{Numeric values} of the exposure
to the environmental stressor referring to the central estimate and (optionally)
to lower and upper bound of the confidence interval. If only one value is provided,
it will be assumed that it refers to population-weighted mean exposure in ug/m3.
If a {vector} is provided, it will be assumed that it refers to the exposure categories
(average exposure in the category) in a exposure distribution
(this information is linked to the proportion of population exposed).}

\item{pop_exp}{\code{Numeric value} or {vector} showing the population exposed for each of the exposure categories. The length of this input variable must be the same as "exp".}

\item{info}{\code{String} or {data frame} showing additional information or id. The suffix "info" will be added to the column name. Default value = NULL.}
}
\value{
TBD. E.g. This function returns a \code{list} with two \code{data.frames}, one with the total health impact and the second one with a row for each category of the exposure distribution.
The data frame include columns such as:
\itemize{
 \item TBD
}
}
\description{
Calculates the health impacts, of an environmental stressor (e.g. noise) using the absolute risk instead of the relative risk
}
\note{
Experimental function
}
\examples{
TBD
}
\author{
Alberto Castro
}
