% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attribute_yld_singlebhd_rr.R
\name{attribute_yld_singlebhd_rr}
\alias{attribute_yld_singlebhd_rr}
\title{Determine years lived with disability (YLD) attributable to the incidence of a specific morbidity health outcome
Calculates the YLDs using a single value for baseline heath data, i.e. without life table. It provides as a result the mean as well as the lower and the higher bounds of the morbidity impact based on the confidence interval of the exposure-response function. Assumption: cases happen at the start of the year.}
\usage{
attribute_yld_singlebhd_rr(
  exp,
  prop_pop_exp = 1,
  cutoff,
  rr,
  erf_increment,
  erf_shape,
  erf_c,
  bhd,
  disability_weight,
  info = NULL
)
}
\arguments{
\item{exp}{\code{Numeric value} showing the population-weighted mean exposure in ug/m3 or \code{vector} showing the exposure category in a exposure distribution (this information is linked to the proportion of population exposed).}

\item{prop_pop_exp}{\code{Numeric value} or \code{Numeric vector} showing the proportion of population exposed (as a fraction, i.e. values between 0 and 1) for a single exposure value or for multiple categories, i.e., a exposure distribution, respectively. If a exposure distribution is used, the dimension of this input must match that of "exp". By default, 1 if a single exposure value is inputted}

\item{cutoff}{\code{Numeric value} showing the cut-off exposure in ug/m3 (i.e. the exposure level below which no health effects occur).}

\item{rr}{\code{Numeric vector} of three numeric values referring to the central estimate as well as the lower and upper bound of the confidence interval.}

\item{erf_increment}{\code{Numeric value} showing the increment of the concentration-response function in ug/m3 (usually 10 or 5).}

\item{erf_shape}{\code{String} to choose among "linear", "log-linear", "linear-log" and "log-log".}

\item{erf_c}{\code{String} showing the user-defined function that puts the relative risk in relation with concentration. The function must have only one variable: c, which means concentration. E.g. "3+c+c^2". Default value = NULL.}

\item{bhd}{\code{Numeric value} showing the baseline health data (incidence of the health outcome in the population).}

\item{disability_weight}{\code{Numeric value} showing the disability weight associated with the morbidity health outcome}

\item{info}{\code{String} showing additional information or id for the pollutant. The suffix "info" will be added to the column name. Default value = NULL.}
}
\value{
This function returns a \code{data.frame} with one row for each value of the
concentration-response function (i.e. central estimate, lower and upper bound confidence interval.
The YLDs are listed in the columns:
\itemize{
 \item yld
 \item yld_rounded
 }
}
\description{
Determine years lived with disability (YLD) attributable to the incidence of a specific morbidity health outcome
Calculates the YLDs using a single value for baseline heath data, i.e. without life table. It provides as a result the mean as well as the lower and the higher bounds of the morbidity impact based on the confidence interval of the exposure-response function. Assumption: cases happen at the start of the year.
}
\note{
Experimental function
}
\examples{
TBD
}
\author{
Axel Luyten
}
