% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attribute_yll_from_lifetable.R
\name{attribute_yll_from_lifetable}
\alias{attribute_yll_from_lifetable}
\title{Attributable years of life lost based on life tables}
\usage{
attribute_yll_from_lifetable(
  exp_central,
  exp_lower = NULL,
  exp_upper = NULL,
  prop_pop_exp = 1,
  cutoff,
  rr_central = NULL,
  rr_lower = NULL,
  rr_upper = NULL,
  erf_increment = NULL,
  erf_shape = NULL,
  erf_c_central = NULL,
  erf_c_lower = NULL,
  erf_c_upper = NULL,
  approach_exposure = "single_year",
  approach_newborns = "without_newborns",
  first_age_pop,
  last_age_pop,
  population_midyear_male,
  population_midyear_female,
  deaths_male = NULL,
  deaths_female = NULL,
  year_of_analysis,
  corrected_discount_rate = NULL,
  min_age = NULL,
  max_age = NULL,
  info = NULL
)
}
\arguments{
\item{exp_central, exp_lower, exp_upper}{\code{Numeric values} of the exposure
to the environmental stressor referring to the central estimate and (optionally)
to lower and upper bound of the confidence interval. If only one value is provided,
it will be assumed that it refers to population-weighted mean exposure in ug/m3.
If a {vector} is provided, it will be assumed that it refers to the exposure categories
(average exposure in the category) in a exposure distribution
(this information is linked to the proportion of population exposed).}

\item{prop_pop_exp}{\code{Numeric value} or {Numeric vector} Fraction (values between 0 & 1) of the total population exposed to (one or more) exposure categories, i.e., a exposure distribution, respectively. If a exposure distribution is used, the dimension of this input variable should be the same as "exp". By default, 1 for single exposure value will be assigned to this input variable assuming a single exposure value, but users can change this value.}

\item{cutoff}{\code{Numeric value} showing the cut-off exposure in ug/m3 (i.e. the exposure level below which no health effects occur).}

\item{rr_central, rr_lower, rr_upper}{Three \code{numeric values} referring to the central estimate of the relative risk and the corresponding lower and upper 95\% confidence interval bounds.}

\item{erf_increment}{\code{Numeric value} showing the increment of the exposure-response function in ug/m3 (usually 10 or 5).}

\item{erf_shape}{\code{String} showing the shape of the exposure-response function to be assumed using the relative risk from the literature as support point. Options: "linear", log_linear", "linear_log", "log_log".}

\item{approach_exposure}{\code{String} showing whether air pollution is constant or only in one year. Options: "single_year" (default), "constant"}

\item{approach_newborns}{\code{String} showing whether newborns are considered in the years after the year of analysis. Options: "without_newborns" (default), "with_newborns"}

\item{first_age_pop}{\code{Numeric value} starting age of the youngest age group from population and life table data (age interval = 1 year)}

\item{last_age_pop}{\code{Numeric value} ending age of the oldest age group from population and life table data (age interval = 1 year)}

\item{population_midyear_male, population_midyear_female}{\code{Numeric vector} containing the mid-year male population for the year of analysis for male and female respectively.}

\item{year_of_analysis}{\code{Numeric value} of the year of analysis, which corresponds to the first year of the life table.}

\item{corrected_discount_rate}{\code{Numeric value} showing the discount rate for future years including correction from inflation rate
estimate and (optionally) the lower bound and the upper bound of the confidence
interval of the baseline health data (e.g. incidence of the health outcome in the population).}

\item{min_age}{\code{Numberic value} of the minimal age to be considered for adults (by default 30, i.e. 30+).}

\item{max_age}{\code{Numberic value} of the maximal age to be considered for infants/children (by default 0, i.e. below 1 year old).}

\item{info}{\code{String} or {data frame} showing additional information or id. The suffix "info" will be added to the column name. Default value = NULL.}
}
\value{
TBD. E.g. This function returns a \code{data.frame} with one row for each value of the
concentration-response function (i.e. central estimate, lower and upper bound confidence interval.
Moreover, the data frame include columns such as:
\itemize{
 \item Attributable fraction
 \item Health impact
 \item Outcome metric
 \item And many more.
}
}
\description{
Calculates the years of life lost attributable to the exposure to an environmental stressor using a life table approach. It provides the central estimate of the impact and the corresponding 95\% confidence intervals (based on the 95\% confidence interval exposure-response function).
}
\note{
Experimental function
}
\examples{
TBD
}
\author{
Alberto Castro
}
