% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bestcost-package.R
\docType{package}
\name{bestcost-package}
\alias{bestcost}
\alias{bestcost-package}
\title{bestcost: Quantification and monetarization of health impacts of environmental stressors.}
\description{
This R package was developed within the framework of the EU project BEST-COST. It aims to quantify and monetize health impacts of environmental stressors focusing on air pollution and noise.
}
\keyword{internal}
