% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_mortality_usingLifetable.R
\name{calculate_mortality_usingLifetable}
\alias{calculate_mortality_usingLifetable}
\title{Calculation of Health Impacts}
\usage{
calculate_mortality_usingLifetable(
  exp,
  cf,
  crf,
  crf_per,
  crf_rescale_method,
  lifetable_withPop_male,
  lifetable_withPop_female,
  firstYear_lifetable,
  nonNatural_death_male,
  nonNatural_death_female,
  pollutant,
  age_group,
  interest_rate,
  increase_of_salary,
  exp_info = NULL,
  cf_info = NULL,
  crf_info = NULL
)
}
\arguments{
\item{exp}{Numeric value showing the population-weighted mean exposure in ug/m3,}

\item{cf}{Numeric value showing the counter-factual scenario (i.e. minimum cut-off concentration) in ug/m3,}

\item{crf}{\code{Vector} of three numeric values referring to the mean as well as the lower bound and upper bound of the confidence interval,}

\item{crf_per}{Numeric value showing the increment of the concentration-response function in ug/m3 (usually 10 or 5),}

\item{crf_rescale_method}{String to choose among "linear" and "loglinear",}

\item{lifetable_withPop_male}{\code{Data frame} with three columns: the first one should refer to age, the second one to the probability of dying and the third one to the population (only male),}

\item{lifetable_withPop_female}{\code{Data frame} with three columns: the first one should refer to age, the second one to the probability of dying and the third one to the population (only female),}

\item{firstYear_lifetable}{Numeric value of the year of analysis, which corresponds to the first year of the life table,}

\item{nonNatural_death_male}{\code{Data frame} with two columns: the first one for age, the second one for the percentage of non-natural deaths (only male),}

\item{nonNatural_death_female}{\code{Data frame} with two columns: the first one for age, the second one for the percentage of non-natural deaths (Only female),}

\item{pollutant}{String with the name of the pollutant,}

\item{age_group}{String with the denomination of the age group (e.g. "adults" or "infants"),}

\item{interest_rate}{Numeric value with the annual interest rate as proportion (i.e. 0.1 instead of 10\%)}

\item{increase_of_salary}{Numeric value with the annual increase of salary as proportion (i.e. 0.1 instead of 10\%)}

\item{exp_info}{\code{Data frame} of one row showing extra information in columns. Default value = NULL,}

\item{cf_info}{\code{Data frame} of one row showing extra information in columns. Default value = NULL,}

\item{crf_info}{\code{Data frame} of three rows (mean, lower bound and upper bound) showing extra information in columns. Default value = NULL,}
}
\value{
This function returns a \code{data.frame} with one row for each value of the
concentration-response function (i.e. mean, lower and upper bound confidence interval.
Moreover, the data frame include columns such as:
\itemize{
\item Attributable fraction
\item Health impact
\item Outcome metric
\item And many more.
}
}
\description{
Calculates the mortality, i.e. premature deaths or years of life lost, attributed to the exposure to an environmental stressor using a life table approach. It provides as a result the mean as well as the lower and the higher bound of the impact based on the confidence interval of the concentration-response function.
}
\note{
Experimental function
}
\examples{
TBD
}
\author{
Alberto Castro
}
