% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_deaths_from_lifetable.R
\name{compare_deaths_from_lifetable}
\alias{compare_deaths_from_lifetable}
\title{Compare attributable deaths obtained applying a life table approach
Calculates the deaths between two scenarios (e.g. before and after a intervention in a health impact assessments). It provides as a result the central estimate as well as the lower and the higher bound of the confidence interval based on the uncertainty of the exposure-response function.}
\usage{
compare_deaths_from_lifetable(
  comparison_method = "delta",
  exp_central_1,
  exp_lower_1 = NULL,
  exp_upper_1 = NULL,
  exp_central_2,
  exp_lower_2 = NULL,
  exp_upper_2 = NULL,
  prop_pop_exp_1 = 1,
  prop_pop_exp_2 = 1,
  cutoff,
  rr_central = NULL,
  rr_lower = NULL,
  rr_upper = NULL,
  erf_increment = NULL,
  erf_shape = NULL,
  erf_c_central = NULL,
  erf_c_lower = NULL,
  erf_c_upper = NULL,
  first_age_pop_1,
  last_age_pop_1,
  prob_natural_death_male_1,
  prob_natural_death_female_1,
  prob_total_death_male_1,
  prob_total_death_female_1,
  population_midyear_male_1,
  population_midyear_female_1,
  year_of_analysis_1,
  first_age_pop_2,
  last_age_pop_2,
  prob_natural_death_male_2,
  prob_natural_death_female_2,
  prob_total_death_male_2,
  prob_total_death_female_2,
  population_midyear_male_2,
  population_midyear_female_2,
  year_of_analysis_2,
  min_age = NULL,
  max_age = NULL,
  info_1 = NULL,
  info_2 = NULL
)
}
\arguments{
\item{comparison_method}{\code{String} showing the method of comparison. Options: "delta" or "pif".}

\item{exp_central_1}{\code{Numeric value} showing the population-weighted mean exposure in ug/m3 or {vector} showing the exposure category in a exposure distribution in the scenario 1.}

\item{exp_central_2}{\code{Numeric value} showing the population-weighted mean exposure in ug/m3 or {vector} showing the exposure category in a exposure distribution in the scenario 2.}

\item{prop_pop_exp_1}{\code{Numeric value} or {vector} showing the proportion of population exposed in the scenario 1. The value is a fraction, i.e. values between 0 and 1) for a single exposure value or for multiple categories, i.e., a exposure distribution, respectively. If a exposure distribution is used, the dimension of this input variable should be the same as "exp". By default, 1 for single exposure value will be assigned to this input variable assuming a single exposure value, but users can change this value.}

\item{prop_pop_exp_2}{\code{Numeric value} or {vector} showing the proportion of population exposed in the scenario 2. The value is a fraction, i.e. values between 0 and 1) for a single exposure value or for multiple categories, i.e., a exposure distribution, respectively. If a exposure distribution is used, the dimension of this input variable should be the same as "exp". By default, 1 for single exposure value will be assigned to this input variable assuming a single exposure value, but users can change this value.}

\item{cutoff}{\code{Numeric value} showing the cut-off exposure in ug/m3 (i.e. the exposure level below which no health effects occur).}

\item{erf_increment}{\code{Numeric value} showing the increment of the concentration-response function in ug/m3 (usually 10 or 5).}

\item{erf_shape}{\code{String} showing the shape of the exposure-response function to be assumed using the relative risk from the literature as support point. Options: "linear", log_linear", "linear_log", "log_log".}

\item{first_age_pop_1}{\code{Numeric value} starting age of the youngest age group from population and life table data in the scenario 1.}

\item{last_age_pop_1}{\code{Numeric value} ending age of the oldest age group from population and life table data in the scenario 1.}

\item{prob_natural_death_female_1}{\code{Numeric vector} containing the probability of dying due to natural cause (excluding non-natural deaths due to violence or accidents) by age or age group for females in the scenario 1.}

\item{prob_total_death_male_1}{\code{Numeric vector} containing the probability of dying due to all causes (including non-natural deaths due to violence or accidents) by age or age group for males in the scenario 1.}

\item{prob_total_death_female_1}{\code{Numeric vector} containing the probability of dying due to all causes (including non-natural deaths due to violence or accidents) by age or age group for females in the scenario 1.}

\item{population_midyear_male_1}{\code{Numeric vector} containing the mid-year male population for the year of analysis of the scenario 1.}

\item{population_midyear_female_1}{\code{Vector} containing the mid-year female population for the year of analysis.}

\item{year_of_analysis_1}{\code{Numeric value} of the year of analysis, which corresponds to the first year of the life table in the scenario 2.}

\item{first_age_pop_2}{\code{Numeric value} starting age of the youngest age group from population and life table data in the scenario 2.}

\item{last_age_pop_2}{\code{Numeric value} ending age of the oldest age group from population and life table data in the scenario 2.}

\item{prob_natural_death_male_2}{\code{Numeric vector} containing the probability of dying due to natural cause (excluding non-natural deaths due to violence or accidents) by age or age group for males in the scenario 2.}

\item{prob_natural_death_female_2}{\code{Numeric vector} containing the probability of dying due to natural cause (excluding non-natural deaths due to violence or accidents) by age or age group for females in the scenario 2.}

\item{prob_total_death_male_2}{\code{Numeric vector} containing the probability of dying due to all causes (including non-natural deaths due to violence or accidents) by age or age group for males in the scenario 2.}

\item{prob_total_death_female_2}{\code{Numeric vector} containing the probability of dying due to all causes (including non-natural deaths due to violence or accidents) by age or age group for females in the scenario 2.}

\item{population_midyear_male_2}{\code{Numeric vector} containing the mid-year male population for the year of analysis of the scenario 2.}

\item{population_midyear_female_2}{\code{Vector} containing the mid-year female population for the year of analysis.}

\item{year_of_analysis_2}{\code{Numeric value} of the year of analysis, which corresponds to the first year of the life table.}

\item{min_age}{\code{Numberic value} of the minimal age to be considered for adults (by default 30, i.e. 30+).}

\item{max_age}{\code{Numberic value} of the maximal age to be considered for infants/children (by default 0, i.e. below 1 year old).}

\item{info_1}{\code{String} or {data frame} showing additional information or id of the scenario 1. The suffix "info" will be added to the column name. Default value = NULL.}

\item{info_2}{\code{String} or {data frame} showing additional information or id of the scenario 1. The suffix "info" will be added to the column name. Default value = NULL.}
}
\value{
TBD. E.g. This function returns a \code{data.frame} with one row for each value of the
concentration-response function (i.e. central estimate, lower and upper bound confidence interval).
Moreover, the data frame includes columns such as:
\itemize{
 \item Attributable fraction
 \item Health impact
 \item Outcome metric
 \item And many more.
}
}
\description{
Compare attributable deaths obtained applying a life table approach
Calculates the deaths between two scenarios (e.g. before and after a intervention in a health impact assessments). It provides as a result the central estimate as well as the lower and the higher bound of the confidence interval based on the uncertainty of the exposure-response function.
}
\note{
Experimental function
}
\examples{
TBD
}
\author{
Alberto Castro
}
