% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_health_ar.R
\name{compare_health_ar}
\alias{compare_health_ar}
\title{Compare attributable health cases based on absolute risk
Calculates the health impacts between two scenarios (e.g. before and after a intervention in a health impact assessments). It provides as a result the central estimate as well as the lower and the higher bound of the confidence interval based on the uncertainty of the exposure-response function.}
\usage{
compare_health_ar(
  exp_1,
  exp_2,
  pop_exp_1,
  pop_exp_2,
  erf_c,
  info_1 = NULL,
  info_2 = NULL
)
}
\arguments{
\item{exp_1}{\code{Numeric value} showing the population-weighted mean exposure in ug/m3 or {vector} showing the exposure category in a exposure distribution in the scenario 1.}

\item{exp_2}{\code{Numeric value} showing the population-weighted mean exposure in ug/m3 or {vector} showing the exposure category in a exposure distribution in the scenario 2.}

\item{pop_exp_1}{\code{Numeric value} or {vector} showing the population exposed for each of the exposure categories in the scenario 1. The length of this input variable must be the same as "exp".}

\item{erf_c}{\code{String} showing the user-defined function that puts the relative risk in relation with concentration. The function must have only one variable: c, which means concentration. E.g. "3+c+c^2". Default value = NULL.}

\item{info_1}{\code{String} or {data frame} showing additional information or id of the scenario 1. The suffix "info" will be added to the column name. Default value = NULL.}

\item{info_2}{\code{String} or {data frame} showing additional information or id of the scenario 1. The suffix "info" will be added to the column name. Default value = NULL.}
}
\value{
TBD. E.g. This function returns a \code{data.frame} with one row for each value of the
concentration-response function (i.e. mean, lower and upper bound confidence interval.
Moreover, the data frame includes columns such as:
\itemize{
 \item Attributable fraction
 \item Health impact
 \item Outcome metric
 \item And many more.
}
}
\description{
Compare attributable health cases based on absolute risk
Calculates the health impacts between two scenarios (e.g. before and after a intervention in a health impact assessments). It provides as a result the central estimate as well as the lower and the higher bound of the confidence interval based on the uncertainty of the exposure-response function.
}
\note{
Experimental function
}
\examples{
TBD
}
\author{
Alberto Castro
}
