% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile_input.R
\name{compile_input}
\alias{compile_input}
\title{Compile input
Compiles the input data of the main function and calculates the population attributable fraction based on the input data (all in one data frame)}
\usage{
compile_input(
  risk_method = NULL,
  health_metric = NULL,
  exp_central,
  exp_lower = NULL,
  exp_upper = NULL,
  prop_pop_exp = NULL,
  pop_exp = NULL,
  cutoff = NULL,
  rr_central,
  rr_lower = NULL,
  rr_upper = NULL,
  erf_increment = NULL,
  erf_shape = NULL,
  erf_c_central = NULL,
  erf_c_lower = NULL,
  erf_c_upper = NULL,
  bhd_central = NULL,
  bhd_lower = NULL,
  bhd_upper = NULL,
  min_age = NULL,
  max_age = NULL,
  geo_id_raw = NULL,
  geo_id_aggregated = NULL,
  info = NULL,
  disability_weight = NULL,
  corrected_discount_rate = NULL,
  duration = NULL
)
}
\arguments{
\item{cutoff}{\code{Numeric value} showing the cut-off exposure in ug/m3 (i.e. the exposure level below which no health effects occur).}

\item{erf_increment}{\code{Numeric value} showing the increment of the concentration-response function in ug/m3 (usually 10 or 5).}

\item{erf_shape}{\code{String} showing the shape of the exposure-response function to be assumed using the relative risk from the literature as support point. Options: "linear", log_linear", "linear_log", "log_log".}

\item{min_age}{\code{Numberic value} of the minimal age to be considered for adults (by default 30, i.e. 30+).}

\item{max_age}{\code{Numberic value} of the maximal age to be considered for infants/children (by default 0, i.e. below 1 year old).}

\item{geo_id_raw}{\code{Vector} showing the id code of the each geographic area considered in the assessment. If a vector is entered here, the data for each geographical area have to be provided as list in the corresponding arguments.}

\item{disability_weight}{\code{Numeric value} showing the disability weight associated with the morbidity health outcome}

\item{corrected_discount_rate}{\code{Numeric value} showing the discount rate for future years including correction from inflation rate}

\item{duration}{\code{Numeric value} showing the disease duration}
}
\value{
This function returns a \code{data.frame} with all input data together
Moreover, the data frame includes columns such as:
\itemize{
 \item Attributable fraction
 \item Health impact
 \item Outcome metric
 \item And many more.
}
}
\description{
Compile input
Compiles the input data of the main function and calculates the population attributable fraction based on the input data (all in one data frame)
}
\note{
Experimental function
}
\examples{
TBD
}
\author{
Alberto Castro
}
\keyword{internal}
