% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile_lifetable_pop.R
\name{compile_lifetable_pop}
\alias{compile_lifetable_pop}
\title{Attributable deaths based on life tables}
\usage{
compile_lifetable_pop(
  first_age_pop,
  last_age_pop,
  prob_natural_death_male,
  prob_natural_death_female,
  prob_total_death_male,
  prob_total_death_female,
  population_midyear_male,
  population_midyear_female,
  geo_id_raw
)
}
\arguments{
\item{first_age_pop}{\code{Numeric value} starting age of the youngest age group from population and life table data}

\item{last_age_pop}{\code{Numeric value} ending age of the oldest age group from population and life table data}

\item{prob_natural_death_male}{\code{Numeric vector} containing the probability of dying due to natural cause (excluding non-natural deaths due to violence or accidents) by age or age group for males.}

\item{prob_natural_death_female}{\code{Numeric vector} containing the probability of dying due to natural cause (excluding non-natural deaths due to violence or accidents) by age or age group for females.}

\item{prob_total_death_male}{\code{Numeric vector} containing the probability of dying due to all causes (including non-natural deaths due to violence or accidents) by age or age group for males.}

\item{prob_total_death_female}{\code{Numeric vector} containing the probability of dying due to all causes (including non-natural deaths due to violence or accidents) by age or age group for females.}

\item{population_midyear_male}{\code{Numeric vector} containing the mid-year male population for the year of analysis.}

\item{population_midyear_female}{\code{Vector} containing the mid-year female population for the year of analysis.}
}
\value{
TBD. E.g. This function returns a \code{data.frame} with one row for each value of the
concentration-response function (i.e. central, lower and upper bound confidence interval.
Moreover, the data frame include columns such as:
\itemize{
 \item Attributable fraction
 \item Health impact
 \item Outcome metric
 \item And many more.
}
}
\description{
Calculates the premature deaths attributable to the exposure to an environmental stressor using a life table approach. It provides the central estimate of the impact and the corresponding 95\% confidence intervals (based on the 95\% confidence interval exposure-response function).
}
\note{
Experimental function
}
\examples{
TBD
}
\author{
Alberto Castro
}
\keyword{internal}
