% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_joining_columns.R
\name{find_joining_columns}
\alias{find_joining_columns}
\title{Find joining columns}
\usage{
find_joining_columns(df1, df2, except = NULL)
}
\arguments{
\item{df1}{First \code{Data frame}}

\item{df2}{Second \code{Data frame}}

\item{except}{\code{Vector} of strings showing columns that have to be excluded although they fulfill the inclusion criteria (same name and value)}
}
\value{
This function returns a \code{vector} of strings
}
\description{
Find columns in two data frames that have same name and identical values (excluding exceptions).- The resulting joining columns are used to dplyr::x_join data frames and add the vector to by=.
}
\note{
Experimental function
}
\examples{
TBD
}
\author{
Alberto Castro & Axel Luyten
}
\keyword{internal}
