% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_daly.R
\name{get_daly}
\alias{get_daly}
\title{Get DALY}
\usage{
get_daly(
  outcome_metric,
  pop_impact,
  year_of_analysis,
  min_age = NULL,
  max_age = NULL,
  input_with_risk_and_pop_fraction,
  corrected_discount_rate = NULL,
  disability_weight = NULL,
  duration = NULL
)
}
\arguments{
\item{outcome_metric}{\code{String} with the metric of the health outcome to be assessed. Options: "deaths", "yll" or "yld".}

\item{pop_impact}{\code{Data frame} with projected population impact over time}

\item{year_of_analysis}{\code{Numeric value} of the year of analysis, which corresponds to the first year of the life table,}

\item{input_with_risk_and_pop_fraction}{\code{Data frame} with meta-information such as input data, additional information and intermediate results.}

\item{corrected_discount_rate}{\code{Numeric value}  with the annual discount rate as proportion (i.e. 0.1 instead of 10\%). It can be calculated as (1+discount_rate_beforeCorrection/1+rate_of_increase)-1}

\item{disability_weight}{\code{Numeric value} showing the disability weight associated with the morbidity health outcome}

\item{duration}{\code{Numeric value} showing the duration (in years) of the morbidity health outcome}

\item{age_min}{\code{Numeric value}  with the minimal age to be considered for adults (by default 30, i.e. 30+),}

\item{age_max}{\code{Numeric value}  with the maximal age to be considered for infants/children (by default 0, i.e. below 1 years old)}

\item{first_age_pop}{\code{Numeric value} starting age of the youngest age group from population and life table data}

\item{last_age_pop}{\code{Numeric value} ending age of the oldest age group from population and life table data}
}
\value{
This function returns a \code{List}
}
\description{
Get attributable disability-adjusted lived years applying a lifetable approach
}
\note{
Experimental function
}
\examples{
TBD
}
\author{
Alberto Castro & Axel Luyten
}
\keyword{internal}
