% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_deaths.R
\name{get_deaths}
\alias{get_deaths}
\title{Get deaths}
\usage{
get_deaths(
  exp,
  cf,
  crf_rescale_method,
  shifted_popOverTime,
  firstYear_lifetable,
  age_group,
  min_age = min_age,
  max_age = max_age
)
}
\arguments{
\item{exp}{Numeric value showing the population-weighted mean exposure in ug/m3,}

\item{cf}{Numeric value showing the counter-factual scenario (i.e. minimum cut-off concentration) in ug/m3,}

\item{crf_rescale_method}{String to choose among "linear" and "loglinear",}

\item{firstYear_lifetable}{Numeric value of the year of analysis, which corresponds to the first year of the life table,}

\item{min_age}{Number with the minimal age to be considered for adults (by default 30, i.e. 30+),}

\item{max_age}{Number with the maximal age to be considered for infants/children (by default 0, i.e. below 1 years old)#' @param age_group String with the denomination of the age group (e.g. "adults" or "infants"),}

\item{shifted_popOvertime}{\code{Data frame} with shifted population over time,}
}
\value{
This function returns a \code{data.frame} with the number of deaths based on the life table
}
\description{
Get deaths
}
\note{
Experimental function
}
\examples{
TBD
}
\author{
Alberto Castro
}
