% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_impact_copy.R
\name{get_impact_copy}
\alias{get_impact_copy}
\title{Attributable health cases based on relative risk}
\usage{
get_impact_copy(
  input,
  lifetable_withPop = NULL,
  year_of_analysis = NULL,
  min_age = NULL,
  max_age = NULL,
  disability_weight = NULL,
  duration = NULL
)
}
\arguments{
\item{input}{\code{Data frame} containing all input data.}
}
\value{
TBD. E.g. This function returns a \code{data.frame} with one row for each value of the
concentration-response function (i.e. central, lower and upper bound confidence interval.
Moreover, the data frame includes columns such as:
\itemize{
 \item Attributable fraction
 \item Health impact
 \item Outcome metric
 \item And many more.
}
}
\description{
Calculates the health impacts for each uncertainty and geo area.
}
\examples{
TBD
}
\author{
Alberto Castro
}
\keyword{internal}
