% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_input_and_paf.R
\name{get_input_and_paf}
\alias{get_input_and_paf}
\title{Get input data and PAF
Compiles the input data of the main function and calculates the population attributable fraction based on the input data (all in one data frame)}
\usage{
get_input_and_paf(
  exp,
  prop_pop_exp,
  cutoff,
  rr,
  rr_increment,
  erf_shape,
  bhd,
  info
)
}
\arguments{
\item{prop_pop_exp}{\code{Numeric value} or {vector} showing the proportion of population exposed (as a fraction, i.e. values between 0 and 1) for a single exposure value or for multiple categories, i.e., a exposure distribution, respectively. If a exposure distribution is used, the dimension of this input variable should be the same as "exp". By default, 1 for single exposure value will be assigned to this input variable assuming a single exposure value, but users can change this value.}

\item{cutoff}{\code{Numeric value} showing the cut-off exposure in ug/m3 (i.e. the exposure level below which no health effects occur).}

\item{rr}{\code{Vector} of three numeric values referring to the mean as well as the lower bound and upper bound of the confidence interval.}

\item{rr_increment}{\code{Numeric value} showing the increment of the concentration-response function in ug/m3 (usually 10 or 5).}

\item{erf_shape}{\code{String} to choose among "linear" and "loglinear".}

\item{bhd}{\code{Numeric value} showing the baseline health data (incidence of the health outcome in the population).}

\item{info}{\code{String} showing additional information or id for the pollutant. The suffix "info" will be added to the column name. Default value = NULL.}
}
\value{
This function returns a \code{data.frame} with one row for each value of the
concentration-response function (i.e. mean, lower and upper bound confidence interval.
Moreover, the data frame includes columns such as:
\itemize{
 \item Attributable fraction
 \item Health impact
 \item Outcome metric
 \item And many more.
}
}
\description{
Get input data and PAF
Compiles the input data of the main function and calculates the population attributable fraction based on the input data (all in one data frame)
}
\note{
Experimental function
}
\examples{
TBD
}
\author{
Alberto Castro
}
