% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_output.R
\name{get_output}
\alias{get_output}
\title{Attributable health cases based on relative risk}
\usage{
get_output(output_raw, method)
}
\arguments{
\item{output_raw}{\code{Data frame} containing all input data and the calculation of health impacts.}

\item{method}{\code{String} showing the calculation method: "relative_risk" or "absolute_risk".}
}
\value{
TBD. E.g. This function returns a \code{data.frame} with one row for each value of the
concentration-response function (i.e. central, lower and upper bound confidence interval.
Moreover, the data frame includes columns such as:
\itemize{
 \item Attributable fraction
 \item Health impact
 \item Outcome metric
 \item And many more.
}
}
\description{
Distributes and store outputs by level of detail by aggregating or filtering impacts.
}
\note{
Experimental function
}
\examples{
TBD
}
\author{
Alberto Castro
}
