% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pop_diff.R
\name{get_pop_diff}
\alias{get_pop_diff}
\title{Get population difference}
\usage{
get_pop_diff(popOverTime_AP, popOverTime_noAP)
}
\arguments{
\item{popOverTime_AP}{\code{Data frame}}

\item{popOverTime_noAP}{\code{Data frame}
#' This function returns a \code{data.frame} containing the difference in population between the reference and the counterfactual scenarios.}
}
\description{
Get population difference
}
\note{
Experimental function
}
\examples{
TBD
}
\author{
Alberto Castro
}
