% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pop_impact.R, R/get_pop_impact_old.R
\name{get_pop_impact}
\alias{get_pop_impact}
\title{Get population impact over time}
\usage{
get_pop_impact(
  lifetable_with_pop,
  year_of_analysis,
  input_with_risk_and_pop_fraction,
  outcome_metric
)

get_pop_impact(
  lifetable_with_pop,
  year_of_analysis,
  input_with_risk_and_pop_fraction,
  outcome_metric
)
}
\arguments{
\item{lifetable_with_pop}{\code{Data frame} with three columns: the first one should refer to age, the second one to the probability of dying and the third one to the population (sex specific),}

\item{year_of_analysis}{\code{Numeric value} of the year of analysis, which corresponds to the first year of the life table}

\item{outcome_metric}{\code{String} to define the outcome metric. Choose between "death", "yll" and "yld"}

\item{pop_fraction}{\code{Data frame} showing the PAF (population attributable fraction) or PIF (population impact fraction) in three rows (central, lower bound and upper bound)}
}
\value{
This function returns a \code{data.frame} with one row for each value of the
concentration-response function (i.e. central estimate, lower and upper bound confidence interval).
Moreover, the data frame include columns such as:
\itemize{
 \item Attributable fraction
 \item Health impact
 \item Outcome metric
 \item And many more.
}

This function returns a \code{data.frame} with one row for each value of the
concentration-response function (i.e. central estimate, lower and upper bound confidence interval).
Moreover, the data frame include columns such as:
\itemize{
 \item Attributable fraction
 \item Health impact
 \item Outcome metric
 \item And many more.
}
}
\description{
Get population impact over time

Get population impact over time
}
\note{
Experimental function

Experimental function
}
\examples{
TBD
TBD
}
\author{
Alberto Castro
}
\keyword{internal}
