% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_popOverTime.R
\name{get_popOverTime}
\alias{get_popOverTime}
\title{Get population over time}
\usage{
get_popOverTime(
  exp,
  cf,
  crf,
  crf_per,
  crf_rescale_method,
  lifetab_withPop,
  firstYear_lifetable,
  nonNatural_death,
  age_group
)
}
\arguments{
\item{exp}{Numeric value showing the population-weighted mean exposure in ug/m3.}

\item{cf}{Numeric value showing the counter-factual scenario (i.e. minimum cut-off concentration) in ug/m3.}

\item{crf}{\code{Vector} of three numeric values referring to the mean as well as the lower bound and upper bound of the confidence interval.}

\item{crf_per}{Numeric value showing the increment of the concentration-response function in ug/m3 (usually 10 or 5)}

\item{crf_rescale_method}{String to choose among "linear" and "loglinear",}

\item{firstYear_lifetable}{Numeric value of the year of analysis, which corresponds to the first year of the life table}

\item{nonNatural_death}{\code{Data frame} with two columns: the first one for age, the second one for the percentage of non-natural deaths (sex specific),}

\item{age_group}{String with the denomination of the age group (e.g. "adults" or "infants"),}

\item{bhd}{Numeric value showing the baseline health data (incidence of the health outcome in the population),}

\item{lifetable_withPop}{\code{Data frame} with three columns: the first one should refer to age, the second one to the probability of dying and the third one to the population (sex specific),}
}
\value{
This function returns a \code{data.frame} the population over time taking into account probability of dying
}
\description{
Get population over time
}
\note{
Experimental function
}
\examples{
TBD
}
\author{
Alberto Castro
}
