% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_popOverTime_noAP.R
\name{get_popOverTime_noAP}
\alias{get_popOverTime_noAP}
\title{Get population over time without changing air pollution exposure}
\usage{
get_popOverTime_noAP(lifetable_wPop, firstYear_lifetable, year_loopStart)
}
\arguments{
\item{lifetable_wPop}{\code{Data frame} with three columns: the first one should refer to age, the second one to the probability of dying and the third one to the population (sex specific),}

\item{firstYear_lifetable}{Numeric value of the year of analysis, which corresponds to the first year of the life table}

\item{year_loopStart}{Numeric value of the year where the deaths have to discounted (normally the first year of the lifetable +1 or +2)}
}
\value{
This function returns a \code{data.frame} with the life table plus an additional column for population one year later without changing air pollution exposure.
}
\description{
Get population over time without changing air pollution exposure
}
\note{
Experimental function
}
\examples{
TBD
}
\author{
Alberto Castro
}
