% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_popOverTime_withAP.R
\name{get_popOverTime_withAP}
\alias{get_popOverTime_withAP}
\title{Get population over time with change in air pollution exposure}
\usage{
get_popOverTime_withAP(
  lifetable_wPop,
  nonNatural_death,
  firstYear_lifetable,
  age_group,
  af
)
}
\arguments{
\item{lifetable_wPop}{\code{Data frame} with three columns: the first one should refer to age, the second one to the probability of dying and the third one to the population (sex specific),}

\item{nonNatural_death}{\code{Data frame} with two columns: the first one for age, the second one for the percentage of non-natural deaths (sex specific),}

\item{firstYear_lifetable}{Numeric value of the year of analysis, which corresponds to the first year of the life table}

\item{age_group}{String with the denomination of the age group (e.g. "adults" or "infants"),}

\item{af}{Attributable fraction}
}
\value{
This function returns a \code{data.frame} with the life table plus an additional column for population one year later without changing air pollution exposure.
}
\description{
Get population over time with change in air pollution exposure
}
\note{
Experimental function
}
\examples{
TBD
}
\author{
Alberto Castro
}
