% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_popSingleYear_withAP.R
\name{get_popSingleYear_withAP}
\alias{get_popSingleYear_withAP}
\title{Get population for single year with change in air pollution exposure}
\usage{
get_popSingleYear_withAP(lifetable_withPop, year_of_analysis, paf)
}
\arguments{
\item{lifetable_withPop}{\code{Data frame} with three columns: the first one should refer to age, the second one to the probability of dying and the third one to the population (sex specific),}

\item{year_of_analysis}{Numeric value of the year of analysis, which corresponds to the first year of the life table}

\item{paf}{Population attributable fraction}
}
\value{
This function returns a \code{data.frame} with the life table plus an additional column for population one year later without changing air pollution exposure.
}
\description{
Get population for single year with change in air pollution exposure
}
\note{
Experimental function
}
\examples{
TBD
}
\author{
Alberto Castro
}
