% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_prob_dying.R
\name{get_prob_dying}
\alias{get_prob_dying}
\title{Get probability of dying}
\arguments{
\item{first_age_pop}{\code{Numeric value} of the first item of the age sequence from population and life table data.}

\item{last_age_pop}{\code{Numeric value} of the last item of the age sequence from population and life table data.}

\item{interval_age_pop}{\code{Vector} containing the interval of the age sequence from population and death data.}

\item{population_midyear}{\code{Vector} containing the mid-year population by age range.}

\item{deaths}{\code{Vector} containing deaths of the year of analysis by age range.}

\item{fraction_of_year_lived}{\code{Numeric value} or \code{Vector} containing the fraction of the age interval that was lived by those who died for each age interval. Default value = 0.5 (i.e. 50%) for all age intervals.}
}
\value{
This function returns a \code{data frame} showing the probability of dying by age interval.
}
\description{
Calculates the probability of dying (natural, non-natural or all cause deaths) based on population and mortality data.
}
\note{
Experimental function
}
\examples{
get_prob_dying(first_age_pop = 0, last_age_pop = 99, interval_age_pop = 1, population = rep(5000, 99), deaths = rep(100, 99))

}
\author{
Alberto Castro
}
