% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_risk.R
\name{get_risk}
\alias{get_risk}
\title{Re-scale the relative risk}
\usage{
get_risk(
  rr = NULL,
  exp,
  cutoff = NULL,
  erf_increment = NULL,
  erf_shape = NULL,
  erf_eq = NULL
)
}
\arguments{
\item{rr}{\code{Numeric vector} containing the relative risk. The data frame must contain the central estimate as well as the lower and upper bound of the exposure-response function.}

\item{exp}{Population exposure to the stressor (e.g. annual population-weighted mean).}

\item{cutoff}{\code{Numeric value} showing the cut-off exposure in ug/m3 (i.e. the exposure level below which no health effects occur).}

\item{erf_increment}{\code{Numeric value} showing the size of the increment in concentration related to the relative risk provided in the literature (e.g. for 10 ug/m3 PM2.5).}

\item{erf_shape}{\code{String} showing the shape of the exposure-response function to be assumed using the relative risk from the literature as support point. Options: "linear", log_linear", "linear_log", "log_log".}

\item{erf_eq}{\code{String} refer to the user-defined function that puts the relative risk in relation with concentration. A string or a data frame can be entered. If a string is entered, it will be converted into a function. The string has to show the function in terms of one variable: "c", which means concentration. E.g. "3+c+c^2". If a data frame is entered, a first column (or with name "c" or "x") should refer to concentration (x-axis) while a second column (or with name "rr" or "y") should refer to relative risk for that concentration (y-axis). A cubic spline natural interpolation will be assumed. Default value = NULL.}
}
\value{
This function returns three \code{values} corresponding to the central estimate as well as the lower and upper bound of the exposure-response function.
}
\description{
Re-scale the relative risk from the increment value in the epidemiological study (e.g. for PM2.5 10 or 5 ug/m3) to the actual population exposure)
}
\note{
Experimental function
}
\examples{
get_risk(rr=1.05, exp=10, cutoff=5, erf_shape="linear" )
}
\author{
Alberto Castro
}
