% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_yld.R
\name{get_yld}
\alias{get_yld}
\title{Get years lived with disability}
\usage{
get_yld(
  pop_impact,
  year_of_analysis,
  min_age,
  max_age,
  meta,
  corrected_discount_rate = 0,
  disability_weight,
  duration
)
}
\arguments{
\item{pop_impact}{\code{Data frame} with projected population impact over time}

\item{year_of_analysis}{\code{Numeric value} of the year of analysis, which corresponds to the first year of the life table,}

\item{meta}{\code{Data frame} with meta-information such as input data, additional information and intermediate results.}

\item{corrected_discount_rate}{\code{Numeric value}  with the annual discount rate as proportion (i.e. 0.1 instead of 10\%). It can be calculated as (1+discount_rate_beforeCorrection/1+rate_of_increase)-1}

\item{duration}{\code{Numeric value} showing the duration (in years) of the morbidity health outcome}

\item{age_min}{\code{Numeric value}  with the minimal age to be considered for adults (by default 30, i.e. 30+),}

\item{age_max}{\code{Numeric value}  with the maximal age to be considered for infants/children (by default 0, i.e. below 1 years old)}

\item{first_age_pop}{\code{Numeric value} starting age of the youngest age group from population and life table data}

\item{last_age_pop}{\code{Numeric value} ending age of the oldest age group from population and life table data}

\item{disbility_weight}{\code{Numeric value} showing the disability weight associated with the morbidity health outcome}
}
\value{
This function returns a \code{List}
}
\description{
Get years lived with disability
}
\note{
Experimental function
}
\examples{
TBD
}
\author{
Axel Luyten
}
\keyword{internal}
