% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_yll.R
\name{get_yll}
\alias{get_yll}
\title{Get years of life lost}
\usage{
get_yll(
  shifted_popOverTime,
  year_of_analysis,
  min_age = min_age,
  max_age = max_age,
  meta,
  corrected_discount_rate
)
}
\arguments{
\item{year_of_analysis}{\code{Numeric value} of the year of analysis, which corresponds to the first year of the life table,}

\item{meta}{\code{Data frame} with meta-information such as input data, additional information and intermediate results.}

\item{corrected_discount_rate}{\code{Numeric value}  with the annual discount rate as proportion (i.e. 0.1 instead of 10\%). It can be calculated as (1+discount_rate_beforeCorrection/1+rate_of_increase)-1}

\item{shifted_popOvertime}{\code{Data frame} with shifted population over time}

\item{age_min}{\code{Numeric value}  with the minimal age to be considered for adults (by default 30, i.e. 30+),}

\item{age_max}{\code{Numeric value}  with the maximal age to be considered for infants/children (by default 0, i.e. below 1 years old)}
}
\value{
This function returns a \code{data.frame} with one row for each value of the
concentration-response function (i.e. mean, lower and upper bound confidence interval.
Moreover, the data frame include columns such as:
\itemize{
 \item Attributable fraction
 \item Health impact
 \item Outcome metric
 \item And many more.
}
}
\description{
Get years of life lost
}
\note{
Experimental function
}
\examples{
TBD
}
\author{
Alberto Castro
}
