% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/include_social.R
\name{include_social}
\alias{include_social}
\title{include_social}
\usage{
include_social(output, deprivation_index)
}
\arguments{
\item{output}{\code{List} produced by \code{bestcost::attribute()} or \code{bestcost::compare()} as results}

\item{deprivation_index}{\code{Numeric value} showing the value of an indicator of economic wealth}
}
\value{
Description of the return value.
}
\description{
Consider socio-economic aspects in the results
}
\examples{
# Example of how to use the function
function_name(param1 = value1, param2 = value2)
}
