% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monetize.R
\name{monetize}
\alias{monetize}
\title{monetize}
\usage{
monetize(impact_raw)
}
\arguments{
\item{impact_raw}{}
}
\value{
Description of the return value.
}
\description{
Monetize health impacts (internal function)
}
\examples{
# Example of how to use the function
function_name(param1 = value1, param2 = value2)
}
