% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/move_rows_up.R
\name{move_rows_up}
\alias{move_rows_up}
\title{Move rows up}
\usage{
move_rows_up(popOTime, year_of_analysis)
}
\arguments{
\item{popOTime}{Population over time,}

\item{year_of_analysis}{Numeric value of the year of analysis, which corresponds to the first year of the life table}
}
\value{
This function returns a shifted value of the population over time taking into account probability of dying
}
\description{
Move rows up
}
\note{
Experimental function
}
\examples{
TBD
}
\author{
Alberto Castro
}
\keyword{internal}
