% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_exposure_data.R
\name{prepare_exposure_data}
\alias{prepare_exposure_data}
\title{Prepare exposure data}
\usage{
prepare_exposure_data(variable_name)
}
\arguments{
\item{erf_ci}{\code{Vector} with the central, lower and upper estimate of the relative risk}
}
\value{
This function returns ...
}
\description{
This function ...
}
\note{
Experimental function
}
\examples{
TBD
}
\author{
Arno Pauwels
}
\keyword{internal}
