% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_pop_copy.R
\name{project_pop_copy}
\alias{project_pop_copy}
\title{Get population over time}
\usage{
project_pop_copy(lifetab_withPop, year_of_analysis, paf, outcome_metric)
}
\arguments{
\item{year_of_analysis}{Numeric value of the year of analysis, which corresponds to the first year of the life table}

\item{paf}{\code{Data frame} with three rows (central estimate as well as lower and upper bound)}

\item{outcome_metric}{\code{String} to define the outcome metric. Choose between "death", "yll" and "yld"}

\item{lifetable_withPop}{\code{Data frame} with three columns: the first one should refer to age, the second one to the probability of dying and the third one to the population (sex specific),}
}
\value{
This function returns a \code{data.frame} the population over time taking into account probability of dying
}
\description{
Get population over time
}
\note{
Experimental function
}
\examples{
TBD
}
\author{
Alberto Castro
}
\keyword{internal}
