% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rescale_rr.R
\name{rescale_rr}
\alias{rescale_rr}
\title{Re-scale the concentration-response function}
\usage{
rescale_rr(rr, exp, cutoff, rr_increment, method)
}
\arguments{
\item{rr}{Data frame containing the concentration-response function as in the epidemiological study, i.e. per the usual concentration increase. The data frame must contain the mean, lower and upper bound of the concentration-response function.}

\item{exp}{Population exposure to the stressor (e.g. annual population-weighted mean).}

\item{cutoff}{\code{Numeric value} showing the cut-off exposure in ug/m3 (i.e. the exposure level below which no health effects occur).}

\item{rr_increment}{Increment of the concentration-response function as in the literature (e.g. for PM2.5 10 or 5 ug/m3).}

\item{method}{Method to re-scale the concentration-response function based on the assumed form of the function. Likely values: "log-linear" or "linear".}
}
\value{
This function returns three \code{values} corresponding to mean, lower bound and upper bound of the concentration-response function.
}
\description{
Re-scale the concentration-response function from the increment value in the epidemiological study (e.g. for PM2.5 10 or 5 ug/m3) to the actual population exposure)
}
\note{
Experimental function
}
\examples{
rescale_rr(rr = pep = 1, metric = "Premature deaths")
}
\author{
Alberto Castro
}
