% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assess_impact_single.R
\name{assess_impact_single}
\alias{assess_impact_single}
\title{Calculation of Health Impacts}
\usage{
assess_impact_single(
  exp,
  cf,
  crf,
  bhd,
  crf_per,
  crf_rescale_method,
  pollutant,
  outcome_metric,
  age_group,
  info_exp = NULL,
  info_cf = NULL,
  info_crf = NULL,
  info_bhd = NULL
)
}
\arguments{
\item{exp}{Numeric value showing the population-weighted mean exposure in ug/m3.}

\item{cf}{Numeric value showing the counter-factual scenario (i.e. minimum cut-off concentration) in ug/m3.}

\item{crf}{\code{Vector} of three numeric values referring to the mean as well as the lower bound and upper bound of the confidence interval.}

\item{bhd}{Numeric value showing the baseline health data (incidence of the health outcome in the population),}

\item{crf_per}{Numeric value showing the increment of the concentration-response function in ug/m3 (usually 10 or 5)}

\item{crf_rescale_method}{String to choose among "linear" and "loglinear",}

\item{pollutant}{String with the name of the pollutant,}

\item{outcome_metric}{String with the name of the health outcome,}

\item{age_group}{String with the denomination of the age group (e.g. "adults" or "infants") ,}

\item{info_exp}{\code{Data frame} of one row showing extra information in columns. Default value = NULL,}

\item{info_cf}{\code{Data frame} of one row showing extra information in columns. Default value = NULL,}

\item{info_crf}{\code{Data frame} of three rows (mean, lower bound and upper bound) showing extra information in columns. Default value = NULL,}

\item{info_bhd}{\code{Data frame} of one row showing extra information in columns. Default value = NULL,}
}
\value{
This function returns a \code{data.frame} with one row for each value of the
concentration-response function (i.e. mean, lower and upper bound confidence interval.
Moreover, the data frame include columns such as:
\itemize{
 \item Attributable fraction
 \item Health impact
 \item Outcome metric
 \item And many more.
}
}
\description{
Calculates the health impacts, mortality or morbidity, of an environmental stressor using a single value for baseline heath data, i.e. without life table. It provides as a result the mean as well as the lower and the higher bound of the impact based on the confidence interval of the concentration-response function.
}
\note{
Experimental function
}
\examples{
TBD
}
\author{
Alberto Castro
}
