% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assess_mortality_lifetable.R
\name{assess_mortality_lifetable}
\alias{assess_mortality_lifetable}
\title{Calculation of Health Impacts}
\usage{
assess_mortality_lifetable(
  exp,
  cf,
  crf,
  crf_per,
  crf_rescale_method,
  first_age_pop,
  last_age_pop,
  interval_age_pop,
  prob_natural_death_male,
  prob_natural_death_female,
  prob_total_death_male,
  prob_total_death_female,
  population_male,
  population_female,
  firstYear_lifetable,
  pollutant,
  age_group,
  corrected_discount_rate,
  min_age = 30,
  max_age = 0,
  exp_info = NULL,
  cf_info = NULL,
  crf_info = NULL
)
}
\arguments{
\item{exp}{Numeric value showing the population-weighted mean exposure in ug/m3,}

\item{cf}{Numeric value showing the counter-factual scenario (i.e. minimum cut-off concentration) in ug/m3,}

\item{crf}{\code{Vector} of three numeric values referring to the mean as well as the lower bound and upper bound of the confidence interval,}

\item{crf_per}{Numeric value showing the increment of the concentration-response function in ug/m3 (usually 10 or 5),}

\item{crf_rescale_method}{String to choose among "linear" and "loglinear",}

\item{first_age_pop}{Numeric value of the first item of the age sequence from population and life table data}

\item{last_age_pop}{Numeric value of the last item of the age sequence from population and life table data}

\item{interval_age_pop}{Numeric value of the interval of the age sequence from population and life table data}

\item{prob_natural_death_male}{\code{Vector} containing the probability of dying because of natural cause (excluding non-natural deaths such as violence or accidents) by age or age group for male,}

\item{prob_natural_death_female}{\code{Vector} containing the probability of dying because of natural cause (excluding non-natural deaths such as violence or accidents) by age or age group for female,}

\item{prob_total_death_male}{\code{Vector} containing the probability of dying because of all causes (including non-natural deaths such as violence or accidents) by age or age group for male,}

\item{prob_total_death_female}{\code{Vector} containing the probability of dying because of all causes (including non-natural deaths such as violence or accidents) by age or age group for female,}

\item{population_male}{\code{Vector} containing the mid-year male population for the year of analysis,}

\item{population_female}{\code{Vector} containing the mid-year female population for the year of analysis,}

\item{firstYear_lifetable}{Numeric value of the year of analysis, which corresponds to the first year of the life table,}

\item{pollutant}{String with the name of the pollutant,}

\item{age_group}{String with the denomination of the age group (e.g. "adults" or "infants"),}

\item{corrected_discount_rate}{Numeric value with the corrected discount rate as proportion (i.e. 0.1 instead of 10\%)}

\item{min_age}{Number with the minimal age to be considered for adults (by default 30, i.e. 30+),}

\item{max_age}{Number with the maximal age to be considered for infants/children (by default 0, i.e. below 1 years old)}

\item{exp_info}{\code{Data frame} of one row showing extra information in columns. Default value = NULL,}

\item{cf_info}{\code{Data frame} of one row showing extra information in columns. Default value = NULL,}

\item{crf_info}{\code{Data frame} of three rows (mean, lower bound and upper bound) showing extra information in columns. Default value = NULL,}
}
\value{
This function returns a \code{data.frame} with one row for each value of the
concentration-response function (i.e. mean, lower and upper bound confidence interval.
Moreover, the data frame include columns such as:
\itemize{
 \item Attributable fraction
 \item Health impact
 \item Outcome metric
 \item And many more.
}
}
\description{
Calculates the mortality, i.e. premature deaths or years of life lost, attributed to the exposure to an environmental stressor using a life table approach. It provides as a result the mean as well as the lower and the higher bound of the impact based on the confidence interval of the concentration-response function.
}
\note{
Experimental function
}
\examples{
TBD
}
\author{
Alberto Castro
}
