% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attribute_health_singlebhd_rr.R
\name{attribute_health_singlebhd_rr}
\alias{attribute_health_singlebhd_rr}
\title{Health impacts based on absolute risk
Calculates the health impacts, mortality or morbidity, of an environmental stressor using a single value for baseline heath data, i.e. without life table. It outputs the central impact estimate as well as the lower and the higher confidence interval bounds, based on the confidence interval of the concentration-response function.}
\usage{
attribute_health_singlebhd_rr(
  exp,
  prop_pop_exp = 1,
  cutoff,
  rr,
  rr_increment,
  erf_shape,
  bhd,
  info_pollutant = NULL,
  info_outcome = NULL,
  info_exp = NULL,
  info_cutoff = NULL,
  info_rr = NULL,
  info_bhd = NULL
)
}
\arguments{
\item{exp}{\code{Numeric value} showing the population-weighted mean exposure in ug/m3 or {vector} showing the exposure category in a exposure distribution (this information is linked to the proportion of population exposed).}

\item{prop_pop_exp}{\code{Numeric value} (if single exposure value) or {Numeric vector} (if exposure distribution with multiple exposure catecories) showing the population exposed as a fraction of the total population (i.e. values between 0 and 1). The dimension of this input variable must be the same as "exp". By default, 1 (i.e. whole population exposed) for single exposure value will be assigned to this input variable.}

\item{cutoff}{\code{Numeric value} () showing the cut-off exposure in ug/m3 (i.e. the exposure level below which there is no risk for the corresponding health outcome).}

\item{rr}{\code{Numeric vector} of three numeric values representing the central relative risk estimate, as well as the lower bound and upper bounds of the confidence interval. Order must be: central estimate, lower CI boud, upper CI bound.}

\item{rr_increment}{\code{Numeric value} showing the increment of the exposure-response function in ug/m3 (usually 10 or 5).}

\item{erf_shape}{\code{String} to choose among "linear" and "loglinear".}

\item{bhd}{\code{Numeric value} showing the baseline health data (incidence of the health outcome in the population).}

\item{info_pollutant}{\code{String} showing additional information or id for the pollutant (added to all rows of the results). Default value = NULL.}

\item{info_outcome}{\code{String} showing additional information or id for the health outcome (added to all rows of the results). Default value = NULL.}

\item{info_exp}{\code{String} showing additional information or id for the exposure (added to all rows of the results). Default value = NULL.}

\item{info_cutoff}{\code{String} showing additional information or id for the cut-off (added to all rows of the results). Default value = NULL.}

\item{info_rr}{\code{String} showing additional information or id for the exposure-response function (added to all rows of the results). Default value = NULL.}

\item{info_bhd}{\code{String} showing additional information or id for the baseline health data (added to all rows of the results). Default value = NULL.}
}
\value{
This function returns a \code{data.frame} with one row for each value of the
concentration-response function (i.e. mean, lower and upper bound confidence interval.
Moreover, the data frame includes columns such as:
\itemize{
 \item Attributable fraction
 \item Health impact
 \item Outcome metric
 \item And many more.
}
}
\description{
Health impacts based on absolute risk
Calculates the health impacts, mortality or morbidity, of an environmental stressor using a single value for baseline heath data, i.e. without life table. It outputs the central impact estimate as well as the lower and the higher confidence interval bounds, based on the confidence interval of the concentration-response function.
}
\note{
Experimental function
}
\examples{
TBD
}
\author{
Alberto Castro
}
