% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_yld_from_prevalence.R
\name{compare_yld_from_prevalence}
\alias{compare_yld_from_prevalence}
\title{Compare attributable health cases based on single baseline health value and relative risk
Calculates the health impacts between two scenarios (e.g. before and after a intervention in a health impact assessments). It provides as a result the central estimate as well as the lower and the higher bound of the confidence interval based on the uncertainty of the exposure-response function.}
\usage{
compare_yld_from_prevalence(
  approach_comparison = "delta",
  exp_central_1,
  exp_lower_1 = NULL,
  exp_upper_1 = NULL,
  exp_central_2,
  exp_lower_2 = NULL,
  exp_upper_2 = NULL,
  prop_pop_exp_1 = 1,
  prop_pop_exp_2 = 1,
  cutoff,
  rr_central = NULL,
  rr_lower = NULL,
  rr_upper = NULL,
  erf_increment = NULL,
  erf_shape = NULL,
  erf_x_central = NULL,
  erf_x_lower = NULL,
  erf_x_upper = NULL,
  bhd_central_1,
  bhd_lower_1 = NULL,
  bhd_upper_1 = NULL,
  bhd_central_2,
  bhd_lower_2 = NULL,
  bhd_upper_2 = NULL,
  dw_central,
  dw_lower = NULL,
  dw_upper = NULL,
  geo_id_raw = NULL,
  geo_id_aggregated = NULL,
  valuation = NULL,
  info_1 = NULL,
  info_2 = NULL
)
}
\arguments{
\item{approach_comparison}{\code{String} showing the method of comparison. Options: "delta" or "pif".}

\item{exp_central_1}{\code{Numeric value} showing the population-weighted mean exposure in ug/m3 or {vector} showing the exposure category in a exposure distribution in the scenario 1.}

\item{exp_central_2}{\code{Numeric value} showing the population-weighted mean exposure in ug/m3 or {vector} showing the exposure category in a exposure distribution in the scenario 2.}

\item{prop_pop_exp_1}{\code{Numeric value} or {vector} showing the proportion of population exposed in the scenario 1. The value is a fraction, i.e. values between 0 and 1) for a single exposure value or for multiple categories, i.e., a exposure distribution, respectively. If a exposure distribution is used, the dimension of this input variable should be the same as "exp". By default, 1 for single exposure value will be assigned to this input variable assuming a single exposure value, but users can change this value.}

\item{prop_pop_exp_2}{\code{Numeric value} or {vector} showing the proportion of population exposed in the scenario 2. The value is a fraction, i.e. values between 0 and 1) for a single exposure value or for multiple categories, i.e., a exposure distribution, respectively. If a exposure distribution is used, the dimension of this input variable should be the same as "exp". By default, 1 for single exposure value will be assigned to this input variable assuming a single exposure value, but users can change this value.}

\item{cutoff}{\code{Numeric value} showing the cut-off exposure in ug/m3 (i.e. the exposure level below which no health effects occur).}

\item{erf_increment}{\code{Numeric value} showing the increment of the concentration-response function in ug/m3 (usually 10 or 5).}

\item{erf_shape}{\code{String} showing the shape of the exposure-response function to be assumed using the relative risk from the literature as support point. Options: "linear", log_linear", "linear_log", "log_log".}

\item{dw_central, dw_lower, dw_upper}{Three \code{Numeric value} showing the disability weights (central estimate, lower and upper 95% confidence intervals) associated with the morbidity health outcome}

\item{geo_id_raw}{\code{Vector} showing the id code of the each geographic area considered in the assessment. If a vector is entered here, the data for each geographical area have to be provided as list in the corresponding arguments.}

\item{info_1}{\code{String} or {data frame} showing additional information or id of the scenario 1. The suffix "info" will be added to the column name. Default value = NULL.}

\item{info_2}{\code{String} or {data frame} showing additional information or id of the scenario 1. The suffix "info" will be added to the column name. Default value = NULL.}
}
\value{
TBD. E.g. This function returns a \code{data.frame} with one row for each value of the
concentration-response function (i.e. central estimate, lower and upper bound confidence interval).
Moreover, the data frame includes columns such as:
\itemize{
 \item Attributable fraction
 \item Health impact
 \item Outcome metric
 \item And many more.
}
}
\description{
Compare attributable health cases based on single baseline health value and relative risk
Calculates the health impacts between two scenarios (e.g. before and after a intervention in a health impact assessments). It provides as a result the central estimate as well as the lower and the higher bound of the confidence interval based on the uncertainty of the exposure-response function.
}
\note{
Experimental function
}
\examples{
TBD
}
\author{
Alberto Castro
}
