% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ci.R
\name{get_ci}
\alias{get_ci}
\title{Get confidence interval}
\usage{
get_ci(
  paf = 0.1,
  bhd_central = 100000,
  bhd_lower = 50000,
  bhd_upper = 150000,
  dw_central = 0.5,
  dw_lower = 0.25,
  dw_upper = 0.75,
  duration_central = 1,
  duration_lower = 0.5,
  duration_upper = 1.5
)
}
\arguments{
\item{paf}{\code{Numerical value} showing the population attributable fraction (paf)}

\item{bhd_central, bhd_lower, bhd_upper}{\code{Numeric value} showing the central estimate, the lower bound and the upper bound of the 95% confidence interval of the baseline health data (bhd) (e.g. prevalence of the health outcome in the population).}

\item{dw_central, dw_lower, dw_upper}{Three \code{Numeric value} showing the disability weights (dw) (central estimate, lower and upper 95% confidence intervals) associated with the morbidity health outcome}

\item{duration_central, duration_lower, duration_upper}{Three \code{Numeric value} showing the duration (central estimate, lower and upper 95% confidence intervals) of the morbidity health outcome}
}
\value{
This function returns a \code{vector} containing the 95% confidence intervals of the inputted YLD impact
}
\description{
Calculate the summary confidence interval of years lived with disability with input sampling
}
\note{
Experimental function
}
\examples{
TBD
}
\author{
Axel Luyten
}
