% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pop_impact.R
\name{get_pop_impact}
\alias{get_pop_impact}
\title{Get population impact over time}
\usage{
get_pop_impact(
  exp,
  cf,
  crf,
  crf_per,
  crf_rescale_method,
  lifetab_withPop,
  nonNatural_death,
  firstYear_lifetable,
  age_group
)
}
\arguments{
\item{exp}{Numeric value showing the population-weighted mean exposure in ug/m3.}

\item{cf}{Numeric value showing the counter-factual scenario (i.e. minimum cut-off concentration) in ug/m3.}

\item{crf}{\code{Vector} of three numeric values referring to the mean as well as the lower bound and upper bound of the confidence interval.}

\item{crf_per}{Numeric value showing the increment of the concentration-response function in ug/m3 (usually 10 or 5)}

\item{crf_rescale_method}{String to choose among "linear" and "loglinear",}

\item{nonNatural_death}{\code{Data frame} with two columns: the first one for age, the second one for the percentage of non-natural deaths (sex specific),}

\item{firstYear_lifetable}{Numeric value of the year of analysis, which corresponds to the first year of the life table}

\item{age_group}{String with the denomination of the age group (e.g. "adults" or "infants"),}

\item{lifetable_withPop_male}{\code{Data frame} with three columns: the first one should refer to age, the second one to the probability of dying and the third one to the population (sex specific),}
}
\value{
This function returns a \code{data.frame} with one row for each value of the
concentration-response function (i.e. mean, lower and upper bound confidence interval.
Moreover, the data frame include columns such as:
\itemize{
 \item Attributable fraction
 \item Health impact
 \item Outcome metric
 \item And many more.
}
}
\description{
Get population impact over time
}
\note{
Experimental function
}
\examples{
TBD
}
\author{
Alberto Castro
}
