% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_popOverTime.R
\name{get_popOverTime}
\alias{get_popOverTime}
\title{Get population over time}
\usage{
get_popOverTime(lifetab_withPop, year_of_analysis, paf)
}
\arguments{
\item{year_of_analysis}{Numeric value of the year of analysis, which corresponds to the first year of the life table}

\item{paf}{\code{Data frame} with three rows (mean, lower bound and upper bound)}

\item{lifetable_withPop}{\code{Data frame} with three columns: the first one should refer to age, the second one to the probability of dying and the third one to the population (sex specific),}
}
\value{
This function returns a \code{data.frame} the population over time taking into account probability of dying
}
\description{
Get population over time
}
\note{
Experimental function
}
\examples{
TBD
}
\author{
Alberto Castro
}
