% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_risk.R
\name{get_risk}
\alias{get_risk}
\title{Re-scale the relative risk}
\usage{
get_risk(
  rr,
  exp,
  cutoff,
  rr_increment,
  erf_shape,
  erf_full = FALSE,
  erf_c = NULL
)
}
\arguments{
\item{rr}{Data frame containing the exposure-response function as in the epidemiological study, i.e. per the usual concentration increase. The data frame must contain the mean, lower and upper bound of the exposure-response function.}

\item{exp}{Population exposure to the stressor (e.g. annual population-weighted mean).}

\item{cutoff}{\code{Numeric value} showing the cut-off exposure in ug/m3 (i.e. the exposure level below which no health effects occur).}

\item{rr_increment}{\code{Numeric value} showing the size of the increment in concentration related to the relative risk provided in the literature (e.g. for 10 ug/m3 PM2.5).}

\item{erf_shape}{\code{String} showing the shape of the exposure-response function to be assumed using the relative risk from the literature as support point. Options: "linear", log_linear", "linear_log", "log_log".}

\item{erf_full}{\code{Boolean value} to show if the exposure-response function is entirely defined by the user  with the argument erf_c (erf_full = TRUE) or by the arguments exp, cutoff, rr_increment and erf_shape (erf_full = TRUE). Default value = FALSE.}

\item{erf_c}{\code{String} showing the user-defined function that puts the relative risk in relation with concentration. The function must have only one variable: c, which means concentration. E.g. "3+c+c^2". Default value = NULL.}
}
\value{
This function returns three \code{values} corresponding to mean, lower bound and upper bound of the exposure-response function.
}
\description{
Re-scale the relative risk from the increment value in the epidemiological study (e.g. for PM2.5 10 or 5 ug/m3) to the actual population exposure)
}
\note{
Experimental function
}
\examples{
get_risk(rr=1.05, exp=10, cutoff=5, erf_shape="linear" )
}
\author{
Alberto Castro
}
