% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_exposure_data.R
\name{prepare_exposure_data}
\alias{prepare_exposure_data}
\title{Prepare exposure data}
\usage{
prepare_exposure_data(poll_grid, geo_units)
}
\arguments{
\item{poll_grid}{\code(SpatRaster) of the pollution data.}

\item{geo_units}{\code(sf) of the geographic units data.}
}
\value{
This function returns a vector of population exposure values.
}
\description{
This function prepares tabular population exposure data compatible with the attribute() and compare() functions, based on spatial pollution and geographic units data.
}
\note{
Experimental function
}
\examples{
TBD
}
\author{
Arno Pauwels
}
\keyword{internal}
